FIX: repository PK Integer로 변경

+++ seas-be/src/main/java/com/ssafy/seas/member/repository/MemberRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.seas.member.entity.Member;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 
 }
